api/requirements.txt
Django>=3.0,<4.0 # Latest version of Django.
django-cors-headers==3.13.0
djangorestframework
pandas
numpy
netcdf4
xarray
